Merge locdesc proposal into spec
This tries to execute the specification changes which are listed in
https://github.com/ccoutant/dwarf-locations/blob/main/005-locations-on-stack.md

Background

The DWARF 5 concept of location descriptions (Section 2.6) limits
their use to cases where the location described is final, and not
subject to some further modification, with two exceptions. First, if
the location description is a memory location description, it is a
simple DWARF expression (Section 2.5) that can be modified by further
DWARF expression operators. Second, for any form of location
description, it can be offset by a fixed number of bits by using a
DW_OP_bit_piece composition operator.

Where do these limitations matter?

Consider the case of a FORTRAN array (as shown in Appendix D in D.2.1)
that has been partially promoted to a register or registers. The
evaluation of its lower and upper bounds depends on the location of
the array as provided by DW_OP_push_object_address. If the array is
not entirely in memory, this operation is not able to provide the
address of the object, as it can only provide a memory address. If
DW_OP_push_object_address were allowed to push a composite location
description on the stack, we could apply further operations to locate
the bounds of the array.

Similarly, consider the case of a pointer-to-member type in C++, where
the object (or part of the object) has been promoted to a register. In
this case, DW_AT_use_location is not able to provide the address of
the object. In optimized code, it sometimes would need to provide a
register location description or a composite location description, but
these cannot be pushed onto the DWARF stack. If DW_AT_use_location
were allowed to push a composite location description on the stack, we
could apply further operations to determine the register location of
the member being referenced.

Also consider the case where a DW_OP_call* operator is used to get the
location of a variable. If the variable happens to be in a register at
the current PC, the call operator cannot succeed, as it cannot push
anything but a memory location on the stack.

All of these cases have a common limiting factor: that location
descriptions cannot be pushed onto the stack, and subsequently
operated on to produce derived location descriptions.  Overview

This proposal removes that limitation. A DWARF expression may evaluate
to either a value or a location. Location descriptions are now simply
DWARF expressions that evaluate to a location.

The DWARF stack is extended so that it can hold elements that are
either (typed) values or (single) locations. The operators in Section
2.6 that previously defined register and implicit locations are now
considered part of a DWARF expression, and are no longer "terminal" in
the sense that they cannot be part of a larger expression.

The literal encoding operations, defined in Section 2.5.1.1, push
values onto the stack, except for DW_OP_addr and DW_OP_addrx, which
push memory locations. These latter two operations are moved to a new
section.

Stack operations, defined in Section 2.5.1.3, can operate on values or
locations, or any combination of the two.

Most existing arithmetic and logical operators, defined in Section
2.5.1.4, continue to be limited to operating on values only.

The DW_OP_deref* and DW_OP_xderef* operators are extended to operate
on any location, and provide the value contained at that location,
whether in memory, in a register, in implicit storage, or a composite
value.

The DW_OP_push_object_address operator pushes a location, which may be
a memory address (as before), or a register, implicit storage, or a
composite.

The DW_AT_use_location attribute provides an expression used to
compute the address of a member for a pointer-to-member type, and
expects the evaluation mechanism to provide the value of the pointer
and the location of the object as implicitly-pushed elements on the
stack. The latter element is now allowed to be any location.

Two new operators, DW_OP_offset and DW_OP_bit_offset, are introduced
that allow a location on the stack to be modified by a byte or a bit
offset.

The composite location operators, DW_OP_piece and DW_OP_bit_piece, are
redefined to build up a composite location, which is held in the top
element of the stack. A new operator, DW_OP_composite, is added to
begin a new (empty) composite location.

The DW_OP_call* operators are now allowed to leave a location on the
stack.

Signed-off-by: Ben Woodard <woodard@redhat.com>