Remove trailing space
The way that I was editing the file introduced considerable trailing
whitespace. Emacs has a very simple command to remove it quickly but
the original document has considerable trailing whitespace itself and
so using the automatic process to remove it makes the diff huge and
makes it less useful. Thus that work was split into a commit by
itself.

Signed-off-by: Ben Woodard <woodard@redhat.com>