Clarifications for Memory Location Descriptions
This is a part in a series of proposals related to support for
debugging on heterogeneous architectures. This proposal lays
groundwork for subsequent proposals by introducing the concept of
addressable memory storage that is associated with each address space
supported by a target architecture, which heretofore had only been
alluded to in the spec (See Section 1.3.2).

BACKGROUND

The changes proposed below introduce the concept of distinct memory
storage for each address space supported by a target architecture, and
a default address space for the target architecture. Subsequent
proposals will make use of this, allowing memory location descriptions
to specify address spaces other than the default one.

Signed-off-by: Ben Woodard <woodard@redhat.com>